Add support for HAL links as array of links
 * Renaming of tests to add space for links as an array
