+++ src/__tests__/linkArrays.js
@@ -1,174 +1,65 @@
 import getHalLink from '..';
 
-describe('HAL link resolver with arrays of links', () => {
-  it('returns URL without params for single non templated link no params are passed', () => {
-    const links = {
-      groups: [
-        {
-          href: 'http://example.com/api/users',
-          templated: false,
-        },
-      ],
-    };
-    expect(getHalLink(links, 'groups')).toBe('http://example.com/api/users');
+describe('HAL link resolver with array of links', () => {
+  it('returns URL without params for single link', () => {
+    const links = [
+      {
+        href: 'http://example.com/api/users/bar',
+        rel: 'edit',
+      },
+      {
+        href: 'http://example.com/api/users/foo',
+        rel: 'delete',
+      },
+    ];
+    expect(getHalLink(links, 'edit')).toBe('http://example.com/api/users/bar');
   });
 
-  it('returns URL without params for single non templated link and params are passed', () => {
-    const links = {
-      groups: [
+  it('returns URL without params for single link', () => {
+    const user = {
+      links: [
         {
-          href: 'http://example.com/api/users',
-          templated: false,
+          href: 'http://example.com/api/users/bar',
+          rel: 'edit',
+        },
+        {
+          href: 'http://example.com/api/users/foo',
+          rel: 'delete',
         },
       ],
     };
-    expect(getHalLink(links, 'groups', {
-      page: 201,
-      size: 13,
-      projection: 'users',
-      filter: 'testFilter',
-      name: 'John',
-    })).toBe('http://example.com/api/users');
+    expect(getHalLink(user, 'edit')).toBe('http://example.com/api/users/bar');
   });
 
-  it('returns URL with translated params for single templated link', () => {
-    const links = {
-      groups: [
-        {
-          href: 'http://example.com/api/users{?projection,page,size}',
-          templated: true,
-        },
-      ],
-    };
-    expect(getHalLink(links, 'groups', {
-      page: 201,
-      size: 13,
-      projection: 'users',
-    })).toBe('http://example.com/api/users?projection=users&page=201&size=13');
+  it('returns null for missing link', () => {
+    const links = [
+      {
+        href: 'http://example.com/api/users',
+        rel: 'edit',
+      },
+      {
+        href: 'http://example.com/api/users',
+        rel: 'delete',
+      },
+    ];
+    expect(getHalLink(links, 'foo')).toBe(null);
   });
 
-  it('returns non templated URL when given templated and non templated link without params', () => {
-    const links = {
-      groups: [
-        {
-          href: 'http://example.com/api/users/foo{?projection,page,size}',
-          templated: true,
-        },
-        {
-          href: 'http://example.com/api/users',
-          templated: false,
-        },
-      ],
-    };
-    expect(getHalLink(links, 'groups')).toBe('http://example.com/api/users');
+  it('returns null for empty array', () => {
+    const links = [];
+    expect(getHalLink(links, 'foo')).toBe(null);
   });
 
-  it('returns URL when given only templated link without params', () => {
-    const links = {
-      groups: [
-        {
-          href: 'http://example.com/api/users/foo{?projection,page,size}',
-          templated: true,
-        },
-      ],
-    };
-    expect(getHalLink(links, 'groups')).toBe('http://example.com/api/users/foo');
-  });
-
-  it('returns first URL when given mutiple templated links without params', () => {
-    const links = {
-      groups: [
-        {
-          href: 'http://example.com/api/users/foo{?projection}',
-          templated: true,
-        },
-        {
-          href: 'http://example.com/api/users/bar{?page,size}',
-          templated: true,
-        },
-      ],
-    };
-    expect(getHalLink(links, 'groups')).toBe('http://example.com/api/users/foo');
-  });
-
-  it('returns URL with options matching params when given mutiple templated links with single param', () => {
-    const links = {
-      groups: [
-        {
-          href: 'http://example.com/api/users/foo{?projection}',
-          templated: true,
-        },
-        {
-          href: 'http://example.com/api/users/bar{?page,size}',
-          templated: true,
-        },
-      ],
-    };
-    expect(getHalLink(links, 'groups', {
-      page: 10,
-    })).toBe('http://example.com/api/users/bar?page=10');
-  });
-
-  it('returns URL with options matching params when given mutiple templated links with two params', () => {
-    const links = {
-      groups: [
-        {
-          href: 'http://example.com/api/users/foo{?projection}',
-          templated: true,
-        },
-        {
-          href: 'http://example.com/api/users/bar{?page,size}',
-          templated: true,
-        },
-      ],
-    };
-    expect(getHalLink(links, 'groups', {
-      page: 10,
-      size: 5,
-    })).toBe('http://example.com/api/users/bar?page=10&size=5');
-  });
-
-  it('returns URL with options matching params when given mutiple templated links with one matching and one non-matching param', () => {
-    const links = {
-      groups: [
-        {
-          href: 'http://example.com/api/users/foo{?projection}',
-          templated: true,
-        },
-        {
-          href: 'http://example.com/api/users/bar{?page,size}',
-          templated: true,
-        },
-      ],
-    };
-    expect(getHalLink(links, 'groups', {
-      page: 10,
-      group: 'users',
-    })).toBe('http://example.com/api/users/bar?page=10');
-  });
-
-  it('returns URL with most options matching params when given mutiple templated links with multiple params', () => {
-    const links = {
-      groups: [
-        {
-          href: 'http://example.com/api/users/foo{?projection}',
-          templated: true,
-        },
-        {
-          href: 'http://example.com/api/users/bar{?projection,page}',
-          templated: true,
-        },
-        {
-          href: 'http://example.com/api/users/xyz{?projection,page,size}',
-          templated: true,
-        },
-      ],
-    };
-    expect(getHalLink(links, 'groups', {
-      page: 10,
-      size: 5,
-      group: 'users',
-      projection: 'groups',
-    })).toBe('http://example.com/api/users/xyz?projection=groups&page=10&size=5');
+  it('returns null for missing href attribute', () => {
+    const links = [
+      {
+        rel: 'edit',
+      },
+      {
+        href: 'http://example.com/api/users',
+        rel: 'delete',
+      },
+    ];
+    expect(getHalLink(links, 'edit')).toBe(null);
   });
 });
